Move CodeGen to it's own file
